requirements.txt
a3m==0.8.1
biopython==1.85
Django==5.1.5
egnn-pytorch==0.2.8
ffindex==0.0.3
ipython==8.28.0
ipywidgets==8.1.5
matplotlib==3.9.2
numpy==1.26.4
paramiko==3.5.1
python-dotenv==1.1.0
requests==2.31.0
scp==0.15.0
torch==2.5.1
tqdm==4.66.2

# WARNING(pigar): some manual fixes might be required as pigar has detected duplicate requirements for the same import name (possibly for different submodules).
# WARNING(pigar): the following duplicate requirements are for the import name: dataset
audio-problems-dataset==0.4
chao-par==0.1
dataset==1.6.2
dataset-manager-py==0.1.1
DatasetRising==1.0.4
db-classifier==0.0.4
encdecmeta==0.0.3
ibotta-uplift==0.0.4
maxda==0.0.1
ml-data-ci==0.0.2
NIA-image-2latex==1.0
nike-detector==0.0.18
pyray2==0.0.4
rgmining-script==0.6.1
stress-testing==0.1
torch-soft==0.1.1
torchblocks-chen==1.0.1.10
wenlan-video==1.0.0
# WARNING(pigar): the following duplicate requirements are for the import name: geometry
aeroterra-ds==0.0.0
cerc-geometry==2.1.3
geometry-packagiing-tomge==1.0.0
my-geometry-lib==0.11
nbody-oross314==0.0.1
PBHnbody==0.0.1
pelatihan-bmkg==0.0.4
pfc-geometry==0.2.16
PyGeometry==1.5.6
PyGeometry-z6==2.1.5
pygeometry-z7==7.2
python-geometry==0.5.0
PyTOUGH==1.6.5
simple-geometry==0.1.4
vesty-bmkg==0.0.3
# WARNING(pigar): the following duplicate requirements are for the import name: model
aicmder==0.5.9
AlgoVision-Quant-Research==0.0.2
arxml-toys==1.0.1
assignment5==1.0
automodel==0.5.9
barbareeka-engine-python==0.0.95
besc==2.0.2
cctv-analysis==0.0.2
celltraj==0.1.65
celltraj-copperma==0.1.0
cnn-interpretation-tool-core==0.0.1
cnn-interpretation-tool-core-by-hyphen==0.0.1
cnn-tool-core==0.0.1
CNN4IE==0.1.9
csx-mortgage-default-prediction-model==0.0.1
dalio==0.0.1
dataanalysistoolkit==1.2.2
dataspine-cli==0.2.5
dataultra.commandlines==0.1
DeepSurrogatepin==1.7
dials==0.0.1
dj-matool==0.1.1
dj-migration-automation==0.1.3
dnac-maps==0.0.2
drl-model==0.1
dxtbx==3.0.4.1
easy-prime==1.2.post2
EL-Status==1.0
encdecmeta==0.0.3
etradebot==3.0.5
fastbrain==0.0.1
fixedwidth-transformer==0.0.1.post1627218962
ga-capstone-hakngrow==0.0.18
gardener-cicd-base==1.2459.0
GCN4LP==0.1
gender-classifier-cnn-usoltsev37==0.0.2
gnn-forecast==0.0.5
GNN4LP==0.1.0
grprec==0.0.2
handObjectacb==0.1.3
hero-auto-sql==1.0.0
hrhr==0.0.3
imm-case-model==1.0.25
ingradient-lib-temp==0.8.1
ingradient-lib-temp2==0.6.61
kiwano-portfolio==1.0.1
KPay-sdk==1.0.1
LibTextClassification==1.3
mallu==0.1.6
Mask-Face-Inference==0.0.1
MiD==0.1.2
mih-naive-bayes==2.0.0
MiMSI==0.4.1
motifcal==0.0.2
mspca==0.0.4
multi-emotion-recognition==0.1.2
nascorp.library==0.1.1
nodeocc==1.0.24
numberAI==0.0.1
package-generator==0.0.1
project-big-data==0.1.0
PROJET-KIT-BIG-DATA==1.0.0
ptmchat==0.0.1
pyactiverecord==0.4.2
pycolonies==1.0.24
pylinktester==0.1
pymongo-model==2.0.2
pyray2==0.0.4
PyVisionTeam17==0.0.5
restartus==2.5.0.1
RNN4IE==0.1.0
roulette-simulator==1.0.0
shaku-database==1.2.23
shaku-database-latest==1.0.10
so==0.5
spinectl==0.2.7
sso-analysis-worker==1.4.1
stDiff-sc==1.1.4
super-taxi-api==0.0.3
superstocks==0.0.1
survival-predictor-model==0.1.0
tid-CICDS-model==0.0.6
tid-my-classification-model==1.0.6
tid-pec2-model==0.0.15
topic-modeling-twitter==0.0.1.2
uleska-automate==0.12
webcrawl==1.1.2
wikipediasummary==1.1.4
wopmars==0.1.5
WpaTestPackage==0.0.2
WTHMDA==0.0.4
xcluster==0.0.1
xjm==0.0.4
Yolo-Distribution-Distillation-Demo==1.0.1
# WARNING(pigar): the following duplicate requirements are for the import name: pdbx
mmcif-pdbx==2.0.1
pdbx==0.3.0
# WARNING(pigar): the following duplicate requirements are for the import name: train
CEFT-Classifier-French==0.1
cupkb==1.0.4
DatasetRising==1.0.4
delphi.ai==0.2.2.0
dl-frame==0.0.0.8
draupnir==0.0.193
dsp-house-prices==0.0.1
dsp-house-prices-muktar==0.1.0
durgamba-housing==0.0.1
example-package-Akalya==0.0.1
HairSys==0.0.12
house-price-prediction==0.0.1
HousePricePrediction-4.1==0.1
housing-5513==0.1
housing-assignment2.1-5505==0.1
housing-library-5490==0.1
housing-library-5491==0.1
housing-library-5492==0.1
housing-library-5493==0.1
housing-library-5494==0.1
housing-library-5495==0.1
housing-library-5496==0.1
housing-library-5497==0.1
housing-library-5503==0.2
housing-library-5509==0.1
housing-library-5510==0.1
housing-library-5511==0.1
housing-library-5512==0.1
housing-library-5514==0.2
housing-library-5515==0.2
housing-library-5520==0.1
housing-price-pred==0.0.3
housing-tiger-2022==0.0.1
housing-traning-5519==0.1
housingLibrary-5508==0.1
nougat-ocr==0.1.17
oShape-package==0.0.1.post7
pytouch==0.4.2
SharingAttention==0.0.1
stockDL==0.2.5
tokyo-jpx==0.0.39
train==0.0.5
train_12306==0.1.6
